+++ demo/js/demo-three.js
@@ -1,24 +1,19 @@
 const React = require('react');
 const ReactDOM = require('react-dom');
-const displace = require('../../dist/displace');
+const displace = require('../../src/displace');
 
 let AppendedToNode = class AppendedToNode extends React.Component {
   render() {
     return (
       <div className="container">
-        <h3>
-          Demo Three's Displaced Element
-        </h3>
+        <h3>Demo Three's Displaced Element</h3>
         <p>
-          I'm appended to a specific node that is passed as a node (not selector string).
+          I'm appended to a specific node that is passed as a node (not selector
+          string).
         </p>
         <p>
-          And I still update correctly.
-          You've clicked "increment demo-three displaced number"
-          {' '}
-          {this.props.number}
-          {' '}
-          time(s).
+          And I still update correctly. You've clicked "increment demo-three
+          displaced number" {this.props.number} time(s).
         </p>
       </div>
     );